requirements.txt
# Core dependencies
python-telegram-bot==13.15  # Using v13 for better compatibility with our current code
openai==1.3.7  # For GPT and DALL-E integration
python-dotenv==1.0.0  # For environment variables management

# Audio processing
pyttsx3==2.90  # For text-to-speech
SpeechRecognition==3.10.0  # For speech recognition
soundfile==0.12.1  # For audio file handling

# Image processing
Pillow==10.1.0  # For image handling

# Utilities
requests==2.31.0  # For HTTP requests
python-json-logger==2.0.7  # For structured JSON logging
tenacity==8.2.3  # For retry mechanisms 